Bug fix regarding the use of ligand annotator
This PR fixes a small logic bug in gen_coords regarding the ligand annotator. Instead of using the processor twice a new instance was generated when trying to split the ligands. That leads to incorrect behavior where the molecule is tagged twice with ligands. This blows up obviously. Tests for this fix are not within polyply but included in the new regression test suite, which will be subject to another PR.